deduplicate global variables
global variables defined in diffirent files with the same name, do clash
at link time. The solution is to define a variable in a single place,
and declare it as extern in the rest. This is of course if sharing
is intended, otherwise the conflicting variables should be renamed diffirently.