bug fixes
Fixing bugs that I have found.
Most are related to converting integers to pointers and vice-versa.
This had the effect that the program would work in 32 bits mode, but not
in 64 bits. Because in 32 bits mode int is the same size as a pointer.
But in 64 bits, pointer is 8 bytes long while int is only 4. This caused
pointers to be corrupted in the process of the conversion, then
the program would crash.